Rebrand marketing page
Inspired by the Bootstrap v5 marketing site.
https://blog.getbootstrap.com/2020/06/16/bootstrap-5-alpha/

This PR upgrades bootstrap to the v5 alpha, removes jQuery.

The landing page refocuses from _what_ Artichoke is to _why_ you would
use or contribute to Artichoke. The audience of the landing page is
shifted toward finding contributors.

The page is much cleaner. Gone are the multicolored boxes. What remains
is three panels:

- Hero
- Quick install instructions
- Project updates / Twitter embeds

The hero has the header:

> Build the next Ruby for Wasm with Artichoke

The value proposition and long-time focus of Artichoke has been to build
a pure-Rust Ruby VM that compiles to the wasm32-unknown-unknown target,
so it makes sense to highlight that in the marketing.

The gradient in the hero and otherwise plain background makes the page
look quite pleasant and not choppy like the colored squares it replaces.