Only use html-loader for partials
This change allows html-webpack-plugin to use the native ejs template
format for the index.html and install.html template files, which allows
using require via html-loader for partials.

These changes allow removing posthtml and a transitive dependency on
markdown-it. This change adds a new custom markdown loader based on
marked.