Remove unused preloads, preload twitter widgets
These warnings appear in the console in prod:

- A preload for 'https://www.artichokeruby.org/main.bundle.js' is found,
  but is not used because the request credentials mode does not match.
  Consider taking a look at crossorigin attribute.
- The resource https://www.artichokeruby.org/main.bundle.js was preloaded
  using link preload but not used within a few seconds from the window's
  load event. Please make sure it has an appropriate `as` value and it is
  preloaded intentionally.

Preloads of the JavaScript bundle are not necessary since this code is
only used for the nav drop downs, which are not render blocking.

Add a preload for Twitter widget JavaScript on the home page since this
JS is render blocking and causes layout shift.