Replace Rust Cargo logo with a libre icon
The acceptability of using the Rust Cargo logo on a non-Rust Project
website is not clear with the changes coming to the Rust Foundation's
trademark policy.

Remove the logo from this repository and replace it with the "boxes"
icon from `bootstrap-icons`:

https://icons.getbootstrap.com/icons/boxes/

The new icon is significantly smaller in bytesize.

This change also updates the nightly moon icon to the latest from the
`bootstrap-icons` project:

https://icons.getbootstrap.com/icons/moon-fill/